Added support for partial upload of a stream
