+++ src/SimpleRestServices/Client/IStringSerializer.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace JSIStudios.SimpleRESTServices.Client
+{
+    public interface IStringSerializer
+    {
+        T Deserialize<T>(string content);
+
+        string Serialize<T>(T obj);
+    }
+}
