+++ src/SimpleRestServices/Core/IRequestLogger.cs
@@ -6,6 +6,6 @@ namespace JSIStudios.SimpleRESTServices.Core
 {
     public interface IRequestLogger
     {
-        void Log(HttpMethod httpMethod, string uri, Dictionary<string, string> requestHeaders, string requestBody, Response response, DateTime requestStartDateUtc, DateTime requestEndDateUtc);
+        void Log(HttpMethod httpMethod, string uri, Dictionary<string, string> requestHeaders, string requestBody, Response response, DateTime requestStartDateUtc, DateTime requestEndDateUtc, Dictionary<string, string> extendedData);
     }
 }
