+++ src/SimpleRestServices/Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.5.2")]
-[assembly: AssemblyFileVersion("1.0.5.2")]
+[assembly: AssemblyVersion("1.0.6.1")]
+[assembly: AssemblyFileVersion("1.0.6.1")]
