+++ src/SimpleRestServices/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Newtonsoft.Json" version="4.5.10" targetFramework="net40" />
+  <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net40" />
 </packages>
